# Cheat to speedup using nqp::iterator+nqp::shift on a list