We can assume the backing array is in a good state so reach in directly during iteration.

`time MVM_SPESH_BLOCKING=1 nqp -e 'my $l := nqp::list(); my int $p := 0; while $p++ < 10_000 { nqp::push($l, $p) }; my int $j := 10000; while $j-- { my $iter := nqp::iterator($l); nqp::while($iter, my $a := nqp::shift($iter)) }'` drops from ~1.46s to ~1.25s and `time MVM_SPESH_BLOCKING=1 nqp -e 'my $l := nqp::list_s(); my int $p := 0; while $p++ < 10_000 { nqp::push_s($l, $p) }; my int $j := 10000; while $j-- { my $iter := nqp::iterator($l); nqp::while($iter, my $a := nqp::shift_s($iter)) }'` drops from ~2.01s to ~1.83s.