@MasterDuke17 2024-05-19
I noticed this when looking at https://github.com/Raku/nqp/issues/685.