@niner 2024-05-19
Why can we assume that the array is in a good state?