@MasterDuke17 2024-05-19
> Why can we assume that the array is in a good state?

Well, I believe we’re allowed to *act* as if it’s in a good state. Modifying an array during iteration is not allowed AIUI. Also, the iterator keeps track of the index compared to the size of the array.