@raiph 2025-07-21
> Modifying an array during iteration is not allowed AIUI.

I couldn't tell if this comment was purely about _nqp_ level commitments or also about _Raku_ level commitments.

For Raku, I'll say this. Shortly after `6.d` shipped, jnthn commented under [an SO answer by Vadim Belman](https://stackoverflow.com/questions/53856578/general-approach-to-modifying-an-array-while-looping-through-it). Based on my existing belief that modification during iteration was fine, plus Vadim's answer, plus jnthn's comment under it, I'd say it looks like it is "officially" allowed in Raku `6.d`.

For nqp, I'll say this. The [doc for `nqp::iterator`](https://github.com/Raku/nqp/blob/4604c6664ea4ba89285137b19dd6533f2ff1f29a/docs/ops.markdown#iterator) says (with my **added EMPHASIS**):

>You can also use `nqp::iterator()` to iterate over a list, which is functionally equivalent to cloning the list and not having to know which form of `shift` to have to use.

```
my $list := nqp::list('a', 'b', 'c');
my $iter := nqp::iterator($list);

while $iter {
    say(nqp::shift($iter));
}
```

> Using `iterator` this way proved to be not very well optimizable, so **this form will probably be DEPRECATED in the future**. Depending on circumstances one could use `clone` instead, or use an index variable...